Split manager domain_ops into entity files
domain_ops.rs held ~80 LibraryManager database-operation methods in one impl
block. Move them verbatim into per-entity child files (release, album, track,
artist, identity, storage, image, import, export) under a domain_ops/ directory.
mod.rs keeps the module doc, the two glob imports, the 7 free helper fns, and a
small impl block for the deletion/cloud-key helpers shared across children.

Children import only `use super::*;`. A descendant module sees domain_ops/mod.rs's
own private `use super::resolve::*;` / `use super::*;` bindings plus the free
helpers, so no per-file import lists are needed.

Placement preserves visibility and keeps every private helper reachable from its
callers. The set_identity chain (resolve_identity_target_album,
fresh_album_for_release, other_release_identities_for_album) stays in identity.rs
with set_identity instead of album.rs, since those private helpers are reachable
only through set_identity. The shared deletion helpers
(queue_release_files_for_deletion, queue_release_cover_for_deletion,
cover_cloud_key) and the pub(super) release_file_cloud_key are hoisted into
mod.rs: they have callers in two child files (and release_file_cloud_key also in
manager/mod.rs), and living in mod.rs keeps pub(super)'s reach identical to the
original single-file module. search_library, which the plan omitted, lands in
album.rs.
