+++ bae-core/src/library/manager/domain_ops/album.rs
@@ -1,0 +1,108 @@
+//! Album domain operations for [`LibraryManager`].
+
+use super::*;
+
+impl LibraryManager {
+    /// Get all albums in the library, sorted by the given criteria.
+    ///
+    /// Pass an empty slice for default sort (newest first).
+    pub async fn get_albums(
+        &self,
+        sort: &[crate::db::AlbumSortCriterion],
+    ) -> Result<Vec<DbAlbum>, LibraryError> {
+        Ok(self.database.get_albums(sort).await?)
+    }
+
+    /// Get a page of albums for lazy loading.
+    pub async fn get_album_page(
+        &self,
+        sort: &[crate::db::AlbumSortCriterion],
+        offset: u64,
+        limit: u64,
+    ) -> Result<Vec<AlbumSummary>, LibraryError> {
+        let raws = self.database.get_album_page(sort, offset, limit).await?;
+        let release_ids: Vec<String> = raws
+            .iter()
+            .flat_map(|r| r.release_ids.iter().cloned())
+            .collect();
+        let covers = self.cover_refs(&release_ids).await?;
+        Ok(raws
+            .into_iter()
+            .map(|raw| resolve_album_summary(raw, |rid| covers.get(rid).cloned()))
+            .collect())
+    }
+
+    /// Count total albums.
+    pub async fn get_album_count(&self) -> Result<u64, LibraryError> {
+        Ok(self.database.get_album_count().await?)
+    }
+
+    /// Get album by ID
+    pub async fn get_album_by_id(&self, album_id: &str) -> Result<Option<DbAlbum>, LibraryError> {
+        Ok(self.database.find_album_by_id(album_id).await?)
+    }
+
+    pub async fn find_album_detail(
+        &self,
+        album_id: &str,
+    ) -> Result<Option<AlbumDetail>, LibraryError> {
+        let Some(raw) = self.database.find_album_detail(album_id).await? else {
+            return Ok(None);
+        };
+        Ok(Some(self.resolve_album_detail(raw).await?))
+    }
+
+    /// Search across albums and tracks.
+    pub async fn search_library(
+        &self,
+        query: &str,
+        limit: usize,
+    ) -> Result<SearchResults, LibraryError> {
+        let raw = self.database.search_library(query, limit).await?;
+        let primary_ids: Vec<String> = raw
+            .albums
+            .iter()
+            .filter_map(|a| a.primary_release_id.clone())
+            .collect();
+        let covers = self.cover_refs(&primary_ids).await?;
+        Ok(resolve_search_results(raw, &covers))
+    }
+
+    /// Delete an album and all its associated data
+    ///
+    /// This will:
+    /// 1. Get all releases for the album
+    /// 2. Queue files for deferred deletion via the pending deletions manifest
+    /// 3. Delete the album from database (cascades to releases and all related data)
+    ///
+    /// File cleanup happens asynchronously via the cleanup service, which retries
+    /// on failure. This prevents orphaned cloud objects when deletion fails.
+    pub async fn delete_album(&self, album_id: &str) -> Result<(), LibraryError> {
+        let releases = self.get_releases_for_album(album_id).await?;
+
+        // Collect track IDs from all releases before deletion for playback cleanup
+        let mut all_track_ids = Vec::new();
+        for release in &releases {
+            if let Ok(tracks) = self.get_tracks(&release.id).await {
+                all_track_ids.extend(tracks.into_iter().map(|t| t.id));
+            }
+            self.queue_release_files_for_deletion(&release.id).await;
+            self.queue_release_cover_for_deletion(&release.id, release.remote)
+                .await;
+        }
+
+        self.database.delete_album(album_id).await?;
+
+        if !all_track_ids.is_empty() {
+            self.emit(LibraryEvent::TracksDeleted {
+                track_ids: all_track_ids,
+            });
+        }
+
+        self.emit_album_removed(album_id, releases.iter().map(|r| r.id.clone()).collect());
+
+        self.spawn_cleanup();
+
+        Ok(())
+    }
+}
