+++ bae-core/src/library/manager/domain_ops/export.rs
@@ -1,0 +1,103 @@
+//! Export domain operations for [`LibraryManager`].
+
+use super::*;
+
+impl LibraryManager {
+    #[cfg(not(any(target_os = "ios", target_os = "android")))]
+    pub async fn export_release(
+        &self,
+        release_id: &str,
+        target_dir: &Path,
+    ) -> Result<(), LibraryError> {
+        ExportService::export_release(release_id, target_dir, self)
+            .await
+            .map_err(LibraryError::Import)
+    }
+
+    /// Assemble everything `ExportService::export_track` needs for a
+    /// track in one pass: source audio bytes, tag fields, cover image path,
+    /// neighbour counts, and the raw audio-format aggregate for decoding.
+    /// Cloud-only tracks download + decrypt here — export never requires a
+    /// local copy.
+    #[cfg(not(any(target_os = "ios", target_os = "android")))]
+    pub async fn get_export_track_plan(
+        &self,
+        track_id: &str,
+    ) -> Result<ExportTrackPlan, LibraryError> {
+        let meta = TrackAudioMeta::resolve(&self.database, track_id).await?;
+
+        let audio_bytes =
+            crate::storage::local::transfer::read_release_file_bytes(&meta.audio_file, self)
+                .await
+                .map_err(|e| {
+                    LibraryError::TrackMapping(format!(
+                        "Couldn't read audio for track {track_id}: {e}"
+                    ))
+                })?;
+
+        let album = self.database.get_album_for_release(&meta.release).await?;
+
+        let album_artists = self.database.get_artists_for_album(&album.id).await?;
+        let artist = join_artist_names(&album_artists);
+
+        let release_tracks = self
+            .database
+            .get_tracks_for_release(&meta.release.id)
+            .await?;
+        let total_tracks = release_tracks.len();
+        let has_multiple_sides = release_tracks
+            .iter()
+            .map(|t| t.side)
+            .collect::<std::collections::HashSet<_>>()
+            .len()
+            > 1;
+        let disc = if has_multiple_sides {
+            Some(meta.track.side)
+        } else {
+            None
+        };
+
+        let year = meta.release.pressing.year.or(album.year);
+
+        let cover_image_bytes = match album.primary_release_id.as_deref() {
+            Some(rid) => self.read_image_blob(rid).await?,
+            None => None,
+        };
+
+        let is_digital =
+            crate::util::format::is_digital_format(meta.release.pressing.format.as_deref());
+
+        let tags = ExportTags {
+            title: meta.track.title.clone(),
+            artist,
+            album: album.title,
+            year,
+            disc,
+        };
+
+        let track_number = meta.track.track_number;
+
+        Ok(ExportTrackPlan {
+            audio_bytes,
+            tags,
+            cover_image_bytes,
+            track_number,
+            total_tracks,
+            is_digital,
+            audio_meta: meta,
+        })
+    }
+
+    #[cfg(not(any(target_os = "ios", target_os = "android")))]
+    pub async fn export_track(
+        &self,
+        track_id: &str,
+        output_path: &Path,
+        format: crate::library::ExportFormat,
+    ) -> Result<(), LibraryError> {
+        let plan = self.get_export_track_plan(track_id).await?;
+        ExportService::export_track(plan, output_path, format)
+            .await
+            .map_err(LibraryError::Import)
+    }
+}
