+++ bae-core/src/library/manager/domain_ops/track.rs
@@ -1,0 +1,150 @@
+//! Track domain operations for [`LibraryManager`].
+
+use super::*;
+
+impl LibraryManager {
+    /// Get tracks for a specific release
+    pub async fn get_tracks(&self, release_id: &str) -> Result<Vec<DbTrack>, LibraryError> {
+        Ok(self.database.get_tracks_for_release(release_id).await?)
+    }
+
+    /// Get ordered track IDs for a release. Use this when the caller only
+    /// needs IDs (queue building, repeat-album rebuild) — avoids pulling
+    /// full `DbTrack` rows.
+    pub async fn get_track_ids(&self, release_id: &str) -> Result<Vec<String>, LibraryError> {
+        Ok(self.database.get_track_ids_for_release(release_id).await?)
+    }
+
+    /// Every track id in the library, in a deterministic base order. Used to
+    /// materialize a `ContextSource::Library` context (shuffle library, and the
+    /// `Context`-repeat re-derive of a library context).
+    pub async fn get_all_track_ids(&self) -> Result<Vec<String>, LibraryError> {
+        Ok(self.database.get_all_track_ids().await?)
+    }
+
+    /// Return the play context for a track: its release id, the release's full
+    /// track order, and the track's index within it. Used by the playback
+    /// service to build the queue around a freshly selected track without
+    /// chaining library calls.
+    pub async fn get_play_context(&self, track_id: &str) -> Result<PlayContext, LibraryError> {
+        let track = self
+            .database
+            .find_track_by_id(track_id)
+            .await?
+            .ok_or_else(|| LibraryError::TrackMapping(format!("Track not found: {}", track_id)))?;
+        let release_id = track.release_id;
+        let track_ids = self.database.get_track_ids_for_release(&release_id).await?;
+        let index = track_ids
+            .iter()
+            .position(|id| id == track_id)
+            .ok_or_else(|| {
+                LibraryError::TrackMapping(format!(
+                    "Track {} not present in its release {}",
+                    track_id, release_id
+                ))
+            })?;
+        Ok(PlayContext {
+            release_id,
+            track_ids,
+            index,
+        })
+    }
+
+    /// Return the subset of `ids` that still exist in the tracks table.
+    /// Used by playback restore to validate a persisted queue in a single
+    /// query instead of one round-trip per track.
+    pub async fn filter_existing_track_ids(
+        &self,
+        ids: &[String],
+    ) -> Result<Vec<String>, LibraryError> {
+        Ok(self.database.filter_existing_track_ids(ids).await?)
+    }
+
+    /// Resolve a list of IDs (which may be album IDs or track IDs) into track IDs.
+    /// Album IDs are expanded to the track IDs of the album's primary release —
+    /// the user's chosen release when set, otherwise the earliest-imported one
+    /// (the fallback `primary_release_id` already encodes).
+    pub async fn resolve_to_track_ids(&self, ids: &[String]) -> Result<Vec<String>, LibraryError> {
+        let mut track_ids = Vec::new();
+        for id in ids {
+            if let Some(detail) = self.find_album_detail(id).await? {
+                if let Some(release) = detail
+                    .releases
+                    .iter()
+                    .find(|r| r.summary.id == detail.primary_release_id)
+                {
+                    track_ids.extend(release.tracks.iter().map(|t| t.id.clone()));
+                }
+            } else {
+                track_ids.push(id.clone());
+            }
+        }
+        Ok(track_ids)
+    }
+
+    pub async fn get_queue_items(
+        &self,
+        entries: &[QueueEntry],
+    ) -> Result<Vec<QueueItem>, LibraryError> {
+        Ok(self.database.get_queue_items(entries).await?)
+    }
+
+    /// Get a specific file by ID
+    ///
+    /// Used during streaming to retrieve the file record after looking up
+    /// the track→file relationship via db_track_position.
+    pub async fn get_file_by_id(&self, file_id: &str) -> Result<Option<DbFile>, LibraryError> {
+        Ok(self.database.find_file_by_id(file_id).await?)
+    }
+
+    /// Get audio format for a track
+    pub async fn get_audio_format_by_track_id(
+        &self,
+        track_id: &str,
+    ) -> Result<Option<DbAudioFormat>, LibraryError> {
+        Ok(self
+            .database
+            .find_audio_format_by_track_id(track_id)
+            .await?)
+    }
+
+    /// Resolve a track's audio into a `ResolvedTrackAudio` with its sample window
+    /// resolved and all raw `Db*` fields hidden.
+    pub async fn resolve_track_audio(
+        &self,
+        track_id: &str,
+    ) -> Result<ResolvedTrackAudio, LibraryError> {
+        let meta = TrackAudioMeta::resolve(&self.database, track_id).await?;
+        Ok(ResolvedTrackAudio::from_meta(&meta))
+    }
+
+    /// Resolve display metadata (artist names, album, cover) for a track at
+    /// playback-preparation time. Done here so `PlaybackService` never sees
+    /// `DbTrack`.
+    pub async fn get_playback_track_info(
+        &self,
+        track_id: &str,
+    ) -> Result<crate::playback::PlaybackTrackInfo, LibraryError> {
+        let track = self
+            .database
+            .find_track_by_id(track_id)
+            .await?
+            .ok_or_else(|| LibraryError::TrackMapping(format!("Track not found: {}", track_id)))?;
+        let release = self.database.get_release_for_track(&track).await?;
+        playback_info_from_track_release(&self.database, &track, &release).await
+    }
+
+    /// Resolve both the audio aggregate and the display metadata for a track in
+    /// a single pass — avoids the `resolve_track_audio` + `get_playback_track_info`
+    /// double-fetch of `DbTrack`/`DbRelease` at playback prep time.
+    pub(crate) async fn resolve_track_audio_and_info(
+        &self,
+        track_id: &str,
+    ) -> Result<(ResolvedTrackAudio, crate::playback::PlaybackTrackInfo), LibraryError> {
+        let meta = TrackAudioMeta::resolve(&self.database, track_id).await?;
+        let audio = ResolvedTrackAudio::from_meta(&meta);
+        let info =
+            playback_info_from_track_release(&self.database, &meta.track, &meta.release).await?;
+        Ok((audio, info))
+    }
+}
